make JS better
