+++ scripts/src/components/datasets-list-english.js
@@ -13,7 +13,8 @@
 import {pick, defaults, filter} from 'lodash'
 
 import TmplDatasetItemEn from '../templates/dataset-item-english'
-import {queryByHook, setContent, createDatasetFilters} from '../util'
+import {queryByHook, setContent, createDatasetFilters, datasetPaginate} from '../util'
+
 
 export default class {
   constructor (opts) {
@@ -30,25 +31,8 @@ export default class {
     const filteredDatasetsEn = filter(opts.datasets, filtersEn)
     const datasetsMarkupEn = filteredDatasetsEn.map(TmplDatasetItemEn)
     setContent(elements.datasetsItemsEnglish, datasetsMarkupEn)
-    $(function() {
-        var pageParts = $("dataset");
-        var numPages = pageParts.length;
-        var perPage = 7;
-        pageParts.slice(perPage).hide();
-        $("#page-nav").pagination({
-            items: numPages,
-            itemsOnPage: perPage,
-            prevText: "Previous",
-            nextText: "Next",
-            cssStyle: "light-theme",
-            onPageClick: function(pageNum) {
-                var start = perPage * (pageNum - 1);
-                var end = start + perPage;
-                pageParts.hide()
-                         .slice(start, end).show();
-            }
-        });
-    });
+    datasetPaginate()
+
 
     // // Dataset count English
     const datasetSuffix =  filteredDatasetsEn.length > 1 ? 's' : ''
@@ -69,25 +53,7 @@ export default class {
       //Dataset Count english
       const resultid = resultsEn.length + ' Datasets'
       setContent(elements.datasetsEnglishCount, resultid)
-      $(function() {
-          var pageParts = $("dataset");
-          var numPages = pageParts.length;
-          var perPage = 7;
-          pageParts.slice(perPage).hide();
-          $("#page-nav").pagination({
-              items: numPages,
-              itemsOnPage: perPage,
-              prevText: "Previous",
-              nextText: "Next",
-              cssStyle: "light-theme",
-              onPageClick: function(pageNum) {
-                  var start = perPage * (pageNum - 1);
-                  var end = start + perPage;
-                  pageParts.hide()
-                           .slice(start, end).show();
-              }
-          });
-      });
+      datasetPaginate()
 
     })
   }
